samples: usb: make WebUSB next sample default
Move legacy WebUSB sample to legacy directory and make WebUSB next
sample default.
