+++ doc/releases/release-notes-4.1.rst
@@ -958,7 +958,7 @@ New Samples
 * :zephyr:code-sample:`uart`
 * :zephyr:code-sample:`usb-midi2-device`
 * :zephyr:code-sample:`usbd-cdc-acm-console`
-* :zephyr:code-sample:`webusb-next`
+* :zephyr:code-sample:`webusb`
 
 Other notable changes
 *********************
