+++ samples/subsys/usb/legacy/webusb/prj.conf
@@ -1,0 +1,13 @@
+CONFIG_STDOUT_CONSOLE=y
+CONFIG_USB_DEVICE_STACK=y
+CONFIG_USB_DEVICE_BOS=y
+CONFIG_SERIAL=y
+CONFIG_USB_DEVICE_PRODUCT="Zephyr WebUSB sample"
+CONFIG_USB_DEVICE_PID=0x000A
+CONFIG_UART_INTERRUPT_DRIVEN=y
+CONFIG_UART_LINE_CTRL=y
+CONFIG_USB_DEVICE_INITIALIZE_AT_BOOT=n
+
+CONFIG_LOG=y
+CONFIG_USB_DRIVER_LOG_LEVEL_ERR=y
+CONFIG_USB_DEVICE_LOG_LEVEL_ERR=y
