+++ samples/subsys/usb/legacy/webusb/sample.yaml
@@ -1,0 +1,7 @@
+sample:
+  name: Web USB
+tests:
+  sample.usb.legacy.webusb:
+    depends_on: usb_device
+    build_only: true
+    tags: usb
