+++ samples/subsys/usb/legacy/webusb/src/webusb.c
