+++ samples/subsys/usb/legacy/webusb/src/webusb.h
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2015-2019 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @file
+ * @brief WebUSB enabled custom class driver header file
+ *
+ * Header file for WebUSB enabled custom class driver
+ */
+
+#ifndef __WEBUSB_SERIAL_H__
+#define __WEBUSB_SERIAL_H__
+
+/**
+ * WebUSB request handlers
+ */
+struct webusb_req_handlers {
+	/* Handler for WebUSB Vendor specific commands */
+	usb_request_handler vendor_handler;
+	/**
+	 * The custom request handler gets a first chance at handling
+	 * the request before it is handed over to the 'chapter 9' request
+	 * handler
+	 */
+	usb_request_handler custom_handler;
+};
+
+/**
+ * @brief Register Custom and Vendor request callbacks
+ *
+ * Function to register Custom and Vendor request callbacks
+ * for handling requests.
+ *
+ * @param [in] handlers Pointer to WebUSB request handlers structure
+ */
+void webusb_register_request_handlers(struct webusb_req_handlers *handlers);
+
+#endif /* __WEBUSB_SERIAL_H__ */
