--- samples/subsys/usb/webusb-next/prj.conf
@@ -1,7 +1,0 @@
-CONFIG_USB_DEVICE_STACK_NEXT=y
-
-CONFIG_LOG=y
-CONFIG_USBD_LOG_LEVEL_WRN=y
-CONFIG_UDC_DRIVER_LOG_LEVEL_WRN=y
-CONFIG_SAMPLE_USBD_PID=0x000A
-CONFIG_SAMPLE_USBD_20_EXTENSION_DESC=y
