--- samples/subsys/usb/webusb-next/sample.yaml
@@ -1,18 +1,0 @@
-sample:
-  name: WebUSB
-tests:
-  sample.usb.webusb-next:
-    depends_on: usbd
-    tags: usb
-    integration_platforms:
-      - nrf52840dk/nrf52840
-      - nrf54h20dk/nrf54h20/cpuapp
-      - frdm_k64f
-      - stm32f723e_disco
-      - nucleo_f413zh
-      - mimxrt685_evk/mimxrt685s/cm33
-      - mimxrt1060_evk/mimxrt1062/qspi
-      - max32690evkit/max32690/m4
-      - samd21_xpro
-      - same54_xpro
-    harness: TBD
