+++ samples/subsys/usb/webusb/CMakeLists.txt
@@ -4,7 +4,6 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(webusb)
 
-target_include_directories(app PRIVATE ${ZEPHYR_BASE}/subsys/usb)
-
+include(${ZEPHYR_BASE}/samples/subsys/usb/common/common.cmake)
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
