+++ samples/subsys/usb/webusb/prj.conf
@@ -1,13 +1,7 @@
-CONFIG_STDOUT_CONSOLE=y
-CONFIG_USB_DEVICE_STACK=y
-CONFIG_USB_DEVICE_BOS=y
-CONFIG_SERIAL=y
-CONFIG_USB_DEVICE_PRODUCT="Zephyr WebUSB sample"
-CONFIG_USB_DEVICE_PID=0x000A
-CONFIG_UART_INTERRUPT_DRIVEN=y
-CONFIG_UART_LINE_CTRL=y
-CONFIG_USB_DEVICE_INITIALIZE_AT_BOOT=n
+CONFIG_USB_DEVICE_STACK_NEXT=y
 
 CONFIG_LOG=y
-CONFIG_USB_DRIVER_LOG_LEVEL_ERR=y
-CONFIG_USB_DEVICE_LOG_LEVEL_ERR=y
+CONFIG_USBD_LOG_LEVEL_WRN=y
+CONFIG_UDC_DRIVER_LOG_LEVEL_WRN=y
+CONFIG_SAMPLE_USBD_PID=0x000A
+CONFIG_SAMPLE_USBD_20_EXTENSION_DESC=y
