+++ samples/subsys/usb/webusb/src/msosv2.h
