+++ samples/subsys/usb/webusb/src/sfunc.c
